Make saa folder
